# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/notifications": {
      "get": {
        "operationId": "NotificationController_getAllNotifications",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all notifications for the current user",
        "tags": [
          "Notification"
        ]
      }
    },
    "/api/notifications/read-all": {
      "patch": {
        "operationId": "NotificationController_markAllAsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark all notifications as read",
        "tags": [
          "Notification"
        ]
      }
    },
    "/api/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a specific notification as read",
        "tags": [
          "Notification"
        ]
      }
    },
    "/api/notifications/{id}/unread": {
      "patch": {
        "operationId": "NotificationController_markAsUnread",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a specific notification as unread",
        "tags": [
          "Notification"
        ]
      }
    },
    "/api/notifications/{id}/delete": {
      "delete": {
        "operationId": "NotificationController_deleteNotification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a specific notification",
        "tags": [
          "Notification"
        ]
      }
    },
...
    "/api/admin/payment-transaction": {
      "get": {
        "operationId": "PaymentTransactionController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all packages",
        "tags": [
          "Payment transaction"
        ]
      }
    },
    "/api/admin/payment-transaction/{id}": {
      "get": {
        "operationId": "PaymentTransactionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one package",
        "tags": [
          "Payment transaction"
        ]
      },
      "delete": {
        "operationId": "PaymentTransactionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payment transaction"
        ]
      }
    },
    "/api/admin/user": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserAdminDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Create a user"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin User Management"
        ]
      },
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [],
        "responses": {
          "default": {
            "description": "Get all users"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin User Management"
        ]
      }
    },
    "/api/admin/user/{id}/approve": {
      "post": {
        "operationId": "UserController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Approve a user"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin User Management"
        ]
      }
    },
    "/api/admin/user/{id}/reject": {
      "post": {
        "operationId": "UserController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Reject a user"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin User Management"
        ]
      }
    },
    "/api/admin/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Get a user by id"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin User Management"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin User Management"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin User Management"
        ]
      }
    },
...
  },
  "info": {
    "title": "Application Name api",
    "description": "Application Name api docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserAdminDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "The email of the user",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "The password of the user",
            "example": "password"
          },
          "type": {
            "type": "string",
            "description": "The type of the user",
            "example": "user"
          },
          "avatar": {
            "type": "string",
            "description": "The avatar of the user",
            "example": "avatar.png"
          },
          "date_of_birth": {
            "type": "string",
            "description": "The date of birth of the user",
            "example": "1990-01-01"
          },
          "gender": {
            "type": "string",
            "description": "The gender of the user",
            "example": "male"
          },
          "address": {
            "type": "string",
            "description": "The location of the user",
            "example": "New York, USA"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "type",
          "avatar",
          "date_of_birth",
          "gender",
          "address"
        ]
      },
      "UpdateUserAdminDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "The email of the user",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "The password of the user",
            "example": "password"
          },
          "type": {
            "type": "string",
            "description": "The type of the user",
            "example": "user"
          },
          "avatar": {
            "type": "string",
            "description": "The avatar of the user",
            "example": "avatar.png"
          },
          "date_of_birth": {
            "type": "string",
            "description": "The date of birth of the user",
            "example": "1990-01-01"
          },
          "gender": {
            "type": "string",
            "description": "The gender of the user",
            "example": "male"
          },
          "address": {
            "type": "string",
            "description": "The location of the user",
            "example": "New York, USA"
          }
        }
      },
...
    }
  }
}